Place users PDF export on User Management instead of profile settings.
Wire Download PDF through AddUserComponent and remove the misplaced export from ProfileSettings admin tab.
